requirements.txt
# Requirements
django==4.0
python-decouple
pillow

# 3rd Party app
djangorestframework==3.13.0
rf-yasg[validation]
django-filter

# Reformation and linting
black
flake8
